Start Postgres in the container
In the Dockerfile we edit the default configuration of Postgres
to change the authentication method for the default user, and create
the core databases. In the setup script we start the Postgres process.

The command to start Postgres was found in the output of the package
manager after installing Postgres.