Move postgres dmp-specific configuration and initialisation from Dockerfile to Python script
This had a number of benefits:
- avoided the need to use an environment variable to pass the postgres username
from the Dockerfile to the setup script for the importing of the test data (see next commits)
- more context given to the operations as one is actually about configuration,
the other ones are effectively the beginning of importing dmp-specific test data into the db
- more consistent with what previous has been done for nginx

The only drawback is adding something like one second to the setup script running time, which is
marginal as the total duration is several minutes. As this stage performance are less important
than having something as simple and easy-to-follow as possible.