Add return type hint to methods
It also rollbacks from a decision of using a fluent interface in one of the classes.
The decision was to either remove it or use it in all classes for consistency.
Given that using a fluent interface was of limited benefit (or it could have
even made the code less readable) I've decided to remove it from that class.