Handle errors that should terminate the script
The method Environment.exit_with_error_message() has been added.
It takes an exception as an input (as opposed to just an error message).
By doing that, the method can print the exception stacktrace, therefore
providing useful debugging information.

Three error conditions have been handled:
- config file is not present
- config file is not parsable as yaml
- a bootstrap command for an app is not available in the config file

The implementation in the last case may look a bit redundant, in that
we raise and catch the same exception. However, I don't feel there isn't
anything wrong with that, as it may be decided to catch the exception
somewhere else in the code at a later date. Moreover, I've tried the
alternative to pass the exception directly to the new method on the lines of
exit_with_error_message(RuntimeException[error_msg]) but the stack trace
wouldn't be available for some reason.