Change the script for updating the repos on demand to use the existing python code
To avoid duplication of logic, the script to update repos on demand uses
the existing Python code.

Unfortunately the script contains a hack to be able to import code from the
env-runner folder. I've marked that with TODO and planning to address that
with the help of somebody else in the team.

While doing this I've also refactored the code to move the responsibility of
managing the repo to a dedicated class, as the Environment class was
starting to look quite busy.