Change the term github-repos to local-repos
I thought that using 'github' was not precise because there isn't anything intrinsic
about Github, so I was thinking of changing it to 'git-repos'. Then I thought
that all our repos are git, so I was about to change it to just 'repos'. But then
I thought it could be confusing whether we mean the remote or local.
So the full thought process was: github-repo -> git-repo -> repo -> local-repo