Add information to script argument
There is no easy way to fix the issue, so just document it should be fine.
I thought of even removing the --dry-run option because of this limitation,
but the option is still useful for getting a good overview of what the setup process
involves, as long as you have the repos cloned, which is quite likely anyway.